Enums moved to dedicated package.
