+++ src/main/java/com/project/airbnb_app/entity/Payment.java
@@ -1,5 +1,6 @@
 package com.project.airbnb_app.entity;
 
+import com.project.airbnb_app.entity.enums.PaymentStatus;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
